+++ CHANGELOG.md
@@ -3982,3 +3982,4 @@
 - [2026-08-28 11:22] test(analytics): add unit tests for analytics business logic
 - [2026-08-28 11:26] feat(test): add bi-directional event bus handler for test
 - [2026-08-28 11:48] docs(theme): update readme and API contracts for theme
+- [2026-08-28 12:06] fix(security): fix memory leak by cancelling security subscription